P1 C4: Move discovery/{memory_gate,system}.rb → inventory/discovery/
Module paths: Legion::LLM::Discovery::{System,MemoryGate} →
Legion::LLM::Inventory::Discovery::{System,MemoryGate}.
Old paths become require-forwarder shims.
router.rb updated to require from new paths.
router/candidates.rb updated to use fully-qualified Inventory::Discovery::MemoryGate.
inventory/discovery.rb updated System refs to new qualified path.
discovery/ directory now contains only forwarder shims (rule_generator.rb stays; deleted P4).
